+++ crates/server/src/error.rs
@@ -40,10 +40,6 @@ pub enum Error {
     #[error(transparent)]
     Stac(#[from] stac::Error),
 
-    /// [stac_api::Error]
-    #[error(transparent)]
-    StacApi(#[from] stac_api::Error),
-
     /// The backend is read-only.
     #[error("this backend is read-only")]
     ReadOnly,